refactor(decrypt): move decryption logic into their own file
